requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-openai
langchain-community
faiss-cpu
